cli: extract onboard cli render support
